update progress bar on cv page

+++ jobSeeker/cvs/upload-cvs.php
@@ -16,17 +16,17 @@ if (isset($_POST['upload'])) {
         if (move_uploaded_file($_FILES['cvs']['tmp_name'], $cvs_path)) {
         } else {
 ?>
-<script>
-alert("Pls check your file")
-</script>
-<?php
+            <script>
+                alert("Pls check your file")
+            </script>
+        <?php
 
         }
     } else {
         ?>
-<script>
-alert("Cvs not Upload sql error")
-</script>
+        <script>
+            alert("Cvs not Upload sql error")
+        </script>
 <?php
     }
 }
@@ -75,27 +75,23 @@ if (isset($_GET['id']) && $_GET['id'] != '' && isset($_GET['action']) && $_GET['
                             $file_path = $data['file_path'];
 
                     ?>
-                    <tr>
-                        <th><?php echo 'cvs/' . htmlspecialchars($id) . '/21' ?></th>
-                        <td><?php echo htmlspecialchars($file_title) ?></td>
-                        <td>
-                            <div class="progress bgl-primary">
-                                <div class="progress-bar" style="width: 70%;" role="progressbar"><span
-                                        class="sr-only">70% Complete</span>
-                                </div>
-                            </div>
-                        </td>
-                        <td>
-                            <div class="d-flex">
-                                <a href="<?php echo $file_path ?>" class="btn btn-primary shadow btn-xs sharp mr-1"><i
-                                        class="fa fa-download"></i></a>
+                            <tr>
+                                <th><?php echo 'cvs/' . htmlspecialchars($id) . '/21' ?></th>
+                                <td><?php echo htmlspecialchars($file_title) ?></td>
+                                <td>
+                                    <div class="progress bgl-primary">
+                                        <div class="progress-bar" style="width: 10%;" role="progressbar"><span class="sr-only">10% Complete</span>
+                                        </div>
+                                    </div>
+                                </td>
+                                <td>
+                                    <div class="d-flex">
+                                        <a href="<?php echo $file_path ?>" class="btn btn-primary shadow btn-xs sharp mr-1"><i class="fa fa-download"></i></a>
 
-                                <a href="my-cv.php?id=<?php echo $id ?>&&action=delete#cvs"
-                                    onClick="return confirm('Are you sure you want to delete this cvs?')"
-                                    class="btn btn-danger shadow btn-xs sharp"><i class="fa fa-trash"></i></a>
-                            </div>
-                        </td>
-                    </tr>
+                                        <a href="my-cv.php?id=<?php echo $id ?>&&action=delete#cvs" onClick="return confirm('Are you sure you want to delete this cvs?')" class="btn btn-danger shadow btn-xs sharp"><i class="fa fa-trash"></i></a>
+                                    </div>
+                                </td>
+                            </tr>
                     <?php
                         }
                     }
@@ -118,21 +114,20 @@ if (isset($_GET['id']) && $_GET['id'] != '' && isset($_GET['action']) && $_GET['
 
 
     ?>
-    <form action="my-cv.php#cvs" method="post" enctype="multipart/form-data" id="upload" class="was-validated">
-        <div class="form-group col-md-6">
-            <label for="cv_title">CVS Title</label>
-            <input type="text" name="cv_title" id="cv_title" class="form-control is-valid" pattern=".([A-zÀ-ž\s]){4,18}"
-                title="4 to 20 characters" required>
-        </div>
-        <div class="form-group col-md-6">
-            <label for="file">Upload Cv</label>
-            <input type="file" name="cvs" id="file" onchange="return fileValidation()" class="form-control" required>
-        </div>
-        <div class="form-group">
-            <button type="submit" class="btn btn-primary" name="upload">Upload</button>
-        </div>
+        <form action="my-cv.php#cvs" method="post" enctype="multipart/form-data" id="upload" class="was-validated">
+            <div class="form-group col-md-6">
+                <label for="cv_title">CVS Title</label>
+                <input type="text" name="cv_title" id="cv_title" class="form-control is-valid" pattern=".([A-zÀ-ž\s]){4,18}" title="4 to 20 characters" required>
+            </div>
+            <div class="form-group col-md-6">
+                <label for="file">Upload Cv</label>
+                <input type="file" name="cvs" id="file" onchange="return fileValidation()" class="form-control" required>
+            </div>
+            <div class="form-group">
+                <button type="submit" class="btn btn-primary" name="upload">Upload</button>
+            </div>
 
-    </form>
+        </form>
     <?php
     } ?>
 
@@ -141,20 +136,20 @@ if (isset($_GET['id']) && $_GET['id'] != '' && isset($_GET['action']) && $_GET['
 </div>
 
 <script>
-function fileValidation() {
-    var fileInput =
-        document.getElementById('file');
+    function fileValidation() {
+        var fileInput =
+            document.getElementById('file');
 
-    var filePath = fileInput.value;
+        var filePath = fileInput.value;
 
-    // Allowing file type
-    var allowedExtensions =
-        /(\.doc|\.docx|\.odt|\.pdf|\.tex|\.txt|\.rtf)$/i;
+        // Allowing file type
+        var allowedExtensions =
+            /(\.doc|\.docx|\.odt|\.pdf|\.tex|\.txt|\.rtf)$/i;
 
-    if (!allowedExtensions.exec(filePath)) {
-        alert('Invalid file type');
-        fileInput.value = '';
-        return false;
+        if (!allowedExtensions.exec(filePath)) {
+            alert('Invalid file type');
+            fileInput.value = '';
+            return false;
+        }
     }
-}
 </script>
